requirements.txt
langchain
openai
huggingface_hub


tiktoken
pypdf


chromadb
pinecone-client
faiss-cpu

sentence_transformers

environs




streamlit
streamlit_chat
